Add `patch` tool as a SDK build dependency
For some dependencies it is more feasible to provide a patch file
to adapt the depdency to FreeOrion needs.  Unfortunately on Windows neither
CMake nor the Windows SDKs provide such a functionality.  Instead the
GnuWin32 `patch.exe` is downloaded on Windows.  On other systems we rely
on the fact that there is an installed version of the `patch` tool
available.