Add `zip` tool as a SDK build dependency
Unfortunately the builtin CMake zip compression tool doesn't allow to
compress whole directories without either prepend every zip entry with
`./` or explicit stating all file on the command line.  Instead the
GnuWin32 `zip.exe` is downloaded on Windows.  On other systems we rely
on the fact that there is an installed version of the `zip` tool
available.