Patch zlib build system
The current upstream build system does not honor the CMake
`BUILD_SHARED_LIBS` variable, allowing users to decide if they want to
build a shared or a static variant of the library.  The patch removes
the zlibstatic target and uses the `BUILD_SHARED_LIBS` variable instead.

The current upstream build system does not install the `program database`
.pdb files, which contain debugging symbols.  The patch install the
`zlib.pdb` file into the `lib` directory of the installation.