Patch libvorbis build system
The current release of libvorbis doesn't provide a CMake build system and
the autotools based one isn't that easy to enable on windows.
Fortunately upstream added a CMake bases build system.  The patch
provides a CMakeLists.txt based on commit `5a263ec` from the xiph/vorbis
github mirror.

The patch also incorporates a fix to link vorbisenc and vorbisfile
with the vorbis library.