Add patches to export boost::{iostreams,locale,log} symbols
When switching the boost dependencies on MacOSX to dynamic libraries
some of those libraries didn't export their symbols properly because
they were not marked as exported by using the `default=hidden` linker
setting.

The patches use the provided boost build features to enable the export
the symbols properly on MacOSX in combination with the Clang compiler.